refactor: update imports to use shadcn directory in attendance pages
